+++ client/app/src/main/java/com/example/witt/di/DataSourceModule.kt
@@ -2,14 +2,14 @@ package com.example.witt.di
 
 import com.example.witt.data.source.local.user_profile.ProfileDataSource
 import com.example.witt.data.source.local.user_profile.ProfileDataSourceImpl
-import com.example.witt.data.source.remote.duplicate_check.DuplicateEmailDataSource
-import com.example.witt.data.source.remote.duplicate_check.DuplicateEmailDataSourceImpl
-import com.example.witt.data.source.remote.memo.EditMemoDataSource
-import com.example.witt.data.source.remote.memo.EditMemoDataSourceImpl
-import com.example.witt.data.source.remote.memo.MakeMemoDataSource
-import com.example.witt.data.source.remote.memo.MakeMemoDataSourceImpl
-import com.example.witt.data.source.remote.plan.detail_plan.AddPlaceDataSource
-import com.example.witt.data.source.remote.plan.detail_plan.AddPlaceDataSourceImpl
+import com.example.witt.data.source.remote.signup.duplicate_check.DuplicateEmailDataSource
+import com.example.witt.data.source.remote.signup.duplicate_check.DuplicateEmailDataSourceImpl
+import com.example.witt.data.source.remote.detail_plan.memo.EditMemoDataSource
+import com.example.witt.data.source.remote.detail_plan.memo.EditMemoDataSourceImpl
+import com.example.witt.data.source.remote.detail_plan.memo.MakeMemoDataSource
+import com.example.witt.data.source.remote.detail_plan.memo.MakeMemoDataSourceImpl
+import com.example.witt.data.source.remote.detail_plan.place.AddPlaceDataSource
+import com.example.witt.data.source.remote.detail_plan.place.AddPlaceDataSourceImpl
 import com.example.witt.data.source.remote.plan.get_plan.GetPlanDataSource
 import com.example.witt.data.source.remote.plan.get_plan.GetPlanDataSourceImpl
 import com.example.witt.data.source.remote.plan.get_plan.GetPlanListDataSource
@@ -20,16 +20,16 @@ import com.example.witt.data.source.remote.plan.join_plan.JoinPlanDataSource
 import com.example.witt.data.source.remote.plan.join_plan.JoinPlanDataSourceImpl
 import com.example.witt.data.source.remote.plan.join_plan.OutPlanDataSource
 import com.example.witt.data.source.remote.plan.join_plan.OutPlanDataSourceImpl
-import com.example.witt.data.source.remote.profile.ProfileUploadDataSource
-import com.example.witt.data.source.remote.profile.ProfileUploadDataSourceImpl
+import com.example.witt.data.source.remote.user.profile.ProfileUploadDataSource
+import com.example.witt.data.source.remote.user.profile.ProfileUploadDataSourceImpl
 import com.example.witt.data.source.remote.signin.SignInDataSource
 import com.example.witt.data.source.remote.signin.SignInDataSourceImpl
 import com.example.witt.data.source.remote.signup.SignUpDataSource
 import com.example.witt.data.source.remote.signup.SignUpDataSourceImpl
-import com.example.witt.data.source.remote.social_signin.SocialSignInDataSource
-import com.example.witt.data.source.remote.social_signin.SocialSignInDataSourceImpl
-import com.example.witt.data.source.remote.token_signin.TokenSignInDataSource
-import com.example.witt.data.source.remote.token_signin.TokenSignInDataSourceImpl
+import com.example.witt.data.source.remote.signin.social_signin.SocialSignInDataSource
+import com.example.witt.data.source.remote.signin.social_signin.SocialSignInDataSourceImpl
+import com.example.witt.data.source.remote.signin.token_signin.TokenSignInDataSource
+import com.example.witt.data.source.remote.signin.token_signin.TokenSignInDataSourceImpl
 import com.example.witt.data.source.remote.user.GetUserInfoDataSource
 import com.example.witt.data.source.remote.user.GetUserInfoDataSourceImpl
 import dagger.Binds
