+++ client/app/src/main/java/com/example/witt/di/RepositoryModule.kt
@@ -6,20 +6,20 @@ import com.example.witt.data.repository.DetailPlanRepositoryImpl
 import com.example.witt.data.repository.PlanRepositoryImpl
 import com.example.witt.data.repository.UserRepositoryImpl
 import com.example.witt.data.source.local.user_profile.ProfileDataSource
-import com.example.witt.data.source.remote.duplicate_check.DuplicateEmailDataSource
-import com.example.witt.data.source.remote.memo.EditMemoDataSource
-import com.example.witt.data.source.remote.memo.MakeMemoDataSource
-import com.example.witt.data.source.remote.plan.detail_plan.AddPlaceDataSource
+import com.example.witt.data.source.remote.signup.duplicate_check.DuplicateEmailDataSource
+import com.example.witt.data.source.remote.detail_plan.memo.EditMemoDataSource
+import com.example.witt.data.source.remote.detail_plan.memo.MakeMemoDataSource
+import com.example.witt.data.source.remote.detail_plan.place.AddPlaceDataSource
 import com.example.witt.data.source.remote.plan.get_plan.GetPlanDataSource
 import com.example.witt.data.source.remote.plan.get_plan.GetPlanListDataSource
 import com.example.witt.data.source.remote.plan.make_plan.MakePlanDataSource
 import com.example.witt.data.source.remote.plan.join_plan.JoinPlanDataSource
 import com.example.witt.data.source.remote.plan.join_plan.OutPlanDataSource
-import com.example.witt.data.source.remote.profile.ProfileUploadDataSource
+import com.example.witt.data.source.remote.user.profile.ProfileUploadDataSource
 import com.example.witt.data.source.remote.signin.SignInDataSource
 import com.example.witt.data.source.remote.signup.SignUpDataSource
-import com.example.witt.data.source.remote.social_signin.SocialSignInDataSource
-import com.example.witt.data.source.remote.token_signin.TokenSignInDataSource
+import com.example.witt.data.source.remote.signin.social_signin.SocialSignInDataSource
+import com.example.witt.data.source.remote.signin.token_signin.TokenSignInDataSource
 import com.example.witt.data.source.remote.user.GetUserInfoDataSource
 import com.example.witt.domain.repository.AuthRepository
 import com.example.witt.domain.repository.DetailPlanRepository
